requirements.txt
python-dotenv
fastapi
fastapi[standard]
uvicorn
pyjwt
pymongo==4.10.1
